+++ src/components/common/Layout/Layout.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 import Head from 'next/head';
 import { css } from '@emotion/react';
-import SideBar from '@/components/organisms/SideBar';
-import Footer from '@/components/organisms/Footer';
+import SideBar from '@/components/common/SideBar';
+import Footer from '@/components/common/Footer';
 
 const Layout: FC = ({ children }) => {
   return (
@@ -13,7 +13,6 @@ const Layout: FC = ({ children }) => {
           name="description"
           content="devChallenges.io - Button component | by h-yoshikawa44"
         />
-        <link rel="icon" href="/favicon.ico" />
       </Head>
       <div css={container}>
         <SideBar />
