+++ src/components/common/SideBar/index.ts
@@ -1,0 +1,3 @@
+import SideBar from './SideBar';
+
+export default SideBar;
