+++ src/components/common/SideBarItem/index.ts
@@ -1,0 +1,3 @@
+import SideBarItem from './SideBarItem';
+
+export default SideBarItem;
