+++ src/components/page/Home/Home.page.tsx
@@ -1,0 +1,12 @@
+import Layout from '@/components/common/Layout';
+import Home from './Home';
+
+const HomePage = () => {
+  return (
+    <Layout>
+      <Home />
+    </Layout>
+  );
+};
+
+export default HomePage;
